Left nav now context-sensitive-able, new favicon, new what's new
Block 'leftnav' now is content block you can rewrite in any template.
Useful for making a block which gives in-page context-sensitive links,
including to internal sections (with #anchorname).
This has been done for the Browse page.

Removing all footer-duplicate noise from leftnav makes it much more
appealing when viewport is compressed. Actually more useful too.

Used a stripes-on-highway image for a favicon, with up-and-left angle
suggesting an upward trajectory. Makes the page much more eye-catching
on browser tabs.

Added a What's New page which gets its info from a
whats_new_items.html page. This way it can be used in multiple places.
Currently, the Home leftnav as well as the About page.